refactor: Move vectorization to its own module
`nodes.rs` is already huge at 2000+ lines and this keeps the vectorization closer to eachother which ought to be easier to read.
